+++ Runtime/Core/Contexts/LangContext.cs
@@ -1,0 +1,13 @@
+namespace Unity.AppUI.Core
+{
+    /// <summary>
+    /// The Lang context of the application.
+    /// </summary>
+    public record LangContext(string lang) : IContext
+    {
+        /// <summary>
+        /// The current language.
+        /// </summary>
+        public string lang { get; } = lang;
+    }
+}