+++ Runtime/Core/Contexts/ThemeContext.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace Unity.AppUI.Core
+{
+    /// <summary>
+    /// The Theme context of the application.
+    /// </summary>
+    public record ThemeContext(string theme) : IContext
+    {
+        /// <summary>
+        /// The current theme.
+        /// </summary>
+        public string theme { get; } = theme;
+    }
+}