+++ Runtime/Core/Contexts/TooltipPlacementContext.cs
@@ -1,0 +1,9 @@
+using Unity.AppUI.UI;
+
+namespace Unity.AppUI.Core
+{
+    public record TooltipPlacementContext(PopoverPlacement placement) : IContext
+    {
+        public PopoverPlacement placement { get; } = placement;
+    }
+}